+++ pyrolite/geochem/__init__.py
@@ -504,11 +504,15 @@ class pyrochem(object):
         norm_to : :class:`str` | :class:`~pyrolite.geochem.norm.Composition` | :class:`numpy.ndarray`
             Which reservoir to normalise REE data to (defaults to :code:`"ChondriteREE_ON"`).
         exclude : :class:`list`, :code:`["Pm", "Eu"]`
-            Which REE elements to exclude from the fit. May wish to include Ce for minerals
+            Which REE elements to exclude from the *fit*. May wish to include Ce for minerals
             in which Ce anomalies are common.
-        params : :class:`list`, :code:`None`
-            Set of predetermined orthagonal polynomial parameters.
-        degree : :class:`int`, 5
+        params : :class:`list` | :class:`str`, :code:`None`
+            Pre-computed parameters for the orthogonal polynomials (a list of tuples).
+            Optionally specified, otherwise defaults the parameterisation as in
+            O'Neill (2016). If a string is supplied, :code:`"O'Neill (2016)"` or
+            similar will give the original defaults, while :code:`"full"` will use all
+            of the REE (including Eu) as a basis for the orthogonal polynomials.
+        degree : :class:`int`, 4
             Maximum degree polynomial fit component to include.
         scale : :class:`str`
             Current units for the REE data, used to scale the reference dataset.