+++ pyrolite/util/lambdas.py
@@ -206,7 +206,7 @@ def get_polynomial_matrix(radii, params=None):
 
 
 @update_docstring_references
-def _lambdas_ONeill2016(df, radii, params=None):
+def lambdas_ONeill2016(df, radii, params=None):
     """
     Implementation of the original algorithm. [#ref_1]_
 
@@ -247,7 +247,7 @@ def _lambdas_ONeill2016(df, radii, params=None):
         columns=[chr(955) + str(d) for d in range(degree)],
         dtype="float32",
     )
-    rad = np.array(radii) # so we can use a boolean index
+    rad = np.array(radii)  # so we can use a boolean index
     for row in range(df.index.size):
         fltr = np.isfinite(df.iloc[row, :].values)  # filter missing
         A = get_polynomial_matrix(rad[fltr], params=params)
@@ -288,7 +288,7 @@ def _lambda_min_func(ls, ys, poly_components, power=1.0):
 
 
 @update_docstring_references
-def _lambdas_optimize(
+def lambdas_optimize(
     df: pd.DataFrame,
     radii,
     params=None,
@@ -353,6 +353,59 @@ def _lambdas_optimize(
 
 
 ########################################################################################
+def _get_params(params=None, degree=4):
+    """
+    Disambiguate parameter specification for orthogonal polynomials.
+
+    Parameters
+    ----------
+    params : :class:`list` | :class:`str`
+        Pre-computed parameters for the orthogonal polynomials (a list of tuples).
+        Optionally specified, otherwise defaults the parameterisation as in
+        O'Neill (2016). [#ref_1]_ If a string is supplied, :code:`"O'Neill (2016)"` or
+        similar will give the original defaults, while :code:`"full"` will use all
+        of the REE (including Eu) as a basis for the orthogonal polynomials.
+    degree : :class:`int`
+        Degree of orthogonal polynomial fit.
+
+    Returns
+    --------
+    params : :class:`list`
+        List of tuples containing a parameterisation of the orthogonal polynomial
+        functions.
+    """
+    if params is None:
+        # use standard parameters as used in O'Neill 2016 paper (exclude Eu)
+        _ree = [i for i in REE() if i not in ["Eu"]]
+        params = orthogonal_polynomial_constants(
+            get_ionic_radii(_ree, charge=3, coordination=8), degree=degree,
+        )
+    elif isinstance(params, str):
+        name = params.replace("'", "").lower()
+        if "full" in name:
+            # use ALL the REE for defininng the orthogonal polynomial functions
+            # (include Eu)
+            _ree = REE()
+        elif ("oneill" in name) and ("2016" in name):
+            # use standard parameters as used in O'Neill 2016 paper (exclude Eu)
+            _ree = [i for i in REE() if i not in ["Eu"]]
+        else:
+            msg = "Parameter specification {} not recognised.".format(params)
+            raise NotImplementedError(msg)
+        print(name, _ree, degree, params)
+        params = orthogonal_polynomial_constants(
+            get_ionic_radii(_ree, charge=3, coordination=8), degree=degree,
+        )
+        print(len(params))
+    else:
+        # check that params is a tuple or list
+        if not isinstance(params, (list, tuple)):
+            msg = "Type {} parameter specification {} not recognised.".format(
+                type(params), params
+            )
+            raise NotImplementedError(msg)
+
+    return params
 
 
 @update_docstring_references
@@ -368,9 +421,12 @@ def calc_lambdas(
     ----------
     df : :class:`pd.DataFrame`
         Dataframe containing REE Data.
-    params : :class:`tuple`
-        Pre-computed parameters for the orthogonal polynomials. Optionally specified,
-        otherwise defaults the parameterisation as in O'Neill (2016). [#ref_1]_
+    params : :class:`list` | :class:`str`
+        Pre-computed parameters for the orthogonal polynomials (a list of tuples).
+        Optionally specified, otherwise defaults the parameterisation as in
+        O'Neill (2016). [#ref_1]_ If a string is supplied, :code:`"O'Neill (2016)"` or
+        similar will give the original defaults, while :code:`"full"` will use all
+        of the REE (including Eu) as a basis for the orthogonal polynomials.
     degree : :class:`int`
         Degree of orthogonal polynomial fit.
     algorithm : :class:`str`
@@ -392,20 +448,20 @@ def calc_lambdas(
            Rare Earth Element Patterns in Basalts. J Petrology 57:1463–1508.
            doi: `10.1093/petrology/egw047 <https://dx.doi.org/10.1093/petrology/egw047>`__
     """
-    # parameters should be set here, and only once
-    if params is None:  # use standard parameters as used in O'Neill 2016 paper
-        default_REE = [i for i in REE() if i not in ["Eu"]]
-        default_radii = get_ionic_radii(default_REE, charge=3, coordination=8)
-        params = orthogonal_polynomial_constants(default_radii, degree=degree)
-    # exclude completely empty columns
+    # parameters should be set here, and only once; these define the inividual
+    # orthogonal polynomial functions which are combined to compose the REE pattern
+    params = _get_params(params=params, degree=degree)
+    # these are the REE which the lambdas will be EVALUATED at; exclude empty columns
     columns = [c for c in df.columns if c not in exclude and np.nansum(df[c])]
-    df[~np.isfinite(df)] = np.nan  # deal with np.nan, np.inf
+    radii = get_ionic_radii(columns, charge=3, coordination=8)
+
     df = df[columns]
-    radii = get_ionic_radii(df.columns.tolist(), charge=3, coordination=8)
+    df[~np.isfinite(df)] = np.nan  # deal with np.nan, np.inf
+
     if "oneill" in algorithm.lower():
-        return _lambdas_ONeill2016(df, radii=radii, params=params, **kwargs)
+        return lambdas_ONeill2016(df, radii=radii, params=params, **kwargs)
     else:
-        return _lambdas_optimize(df, radii=radii, params=params, **kwargs)
+        return lambdas_optimize(df, radii=radii, params=params, **kwargs)
 
 
 def plot_lambdas_components(lambdas, ax=None, params=None, degree=4, **kwargs):