+++ test/geochem/geochem_transform.py
@@ -3,6 +3,7 @@ import numpy as np
 from pyrolite.util.synthetic import test_df, test_ser
 from pyrolite.geochem.transform import *
 from pyrolite.geochem.norm import get_reference_composition
+from pyrolite.util.lambdas import orthogonal_polynomial_constants
 
 
 class TestToMolecular(unittest.TestCase):
@@ -496,6 +497,27 @@ class TestLambdaLnREE(unittest.TestCase):
         ]:
             with self.subTest(norm_to=norm_to):
                 ret = lambda_lnREE(self.df, norm_to=norm_to, degree=self.default_degree)
+                self.assertTrue(ret.columns.size == self.default_degree)
+
+    def test_params(self):
+        # the first three all have the same result - defaulting to ONeill 2016
+        # the two following use a full REE set for the OP basis function definition
+        # the last illustrates that custom parameters could be used
+        degree = self.default_degree
+        for params in [
+            None,
+            "ONeill2016",
+            "O'Neill (2016)",
+            "full",
+            "Full",
+            orthogonal_polynomial_constants(
+                get_ionic_radii(REE(), charge=3, coordination=8),  # xs
+                degree=self.default_degree,
+            ),
+        ]:
+            with self.subTest(params=params):
+                ret = lambda_lnREE(self.df, params=params, degree=degree)
+                print(degree, params, ret.columns)
                 self.assertTrue(ret.columns.size == self.default_degree)